Reorganization of files to match project format
