+++ ecommerce-frontend/src/App.js
@@ -1,7 +1,7 @@
 import logo from './logo.svg';
-import './App.css';
-import Login from './login/login/login';
-import Signup from './login/signup/signup';
+//import './App.css';
+import Login from './components/login';
+import Signup from './components/signup';
 
 
 import './styles/App.scss';
